moive FSBCompatible trait to mod
